Feat(Month1: Refinement - adding a dual screen setup to have a better work lab

+++ ResearchProject_CPE800/NVMe_researchproj/setup.sh
@@ -8,65 +8,74 @@ TP_DIR="$(pwd)/thirdparties/non-permanent"
 ROOTFS="$TP_DIR/rootfs"
 
 echo "--- 1. Cleaning up thirdparties ---"
-# Unmount if necessary to avoid 'Device busy'
 sudo umount -l $ROOTFS/dev $ROOTFS/proc $ROOTFS/sys 2>/dev/null || true
 
 echo "--- 2. Building BusyBox in $TP_DIR ---"
 cd $TP_DIR
+rm -rdf rootfs/
 
 if [ ! -d "busybox-$BB_VER" ]; then
     wget -qO- https://busybox.net/downloads/busybox-$BB_VER.tar.bz2 | tar -xjf -
-    mv busybox-$BB_VER busybox_src
 fi
 
-cd busybox_src
-# Force static linking so we don't need shared libs in the initrd
-make ARCH=riscv CROSS_COMPILE=riscv64-linux-gnu- -j$(nproc)
+cd busybox-$BB_VER
+sed -i 's/# CONFIG_STATIC is not set/CONFIG_STATIC=y/' .config
+sed -i 's/CONFIG_PIE=y/# CONFIG_PIE is not set/' .config
+sed -i 's/CONFIG_TC=y/# CONFIG_TC is not set/' .config
+
+make ARCH=riscv CROSS_COMPILE=$CROSS -j$(nproc)
 make ARCH=$ARCH CROSS_COMPILE=$CROSS -j$(nproc) install CONFIG_PREFIX=$ROOTFS
-     
 cd ../..
 
 echo "--- 3. Creating Initrd with Folder Sharing ---"
-mkdir -p $ROOTFS/dev
-mkdir -p $ROOTFS/proc
-mkdir -p $ROOTFS/sys
-mkdir -p $ROOTFS/etc
-mkdir -p $ROOTFS/src
+mkdir -p $ROOTFS/dev $ROOTFS/proc $ROOTFS/sys $ROOTFS/etc $ROOTFS/src $ROOTFS/lib
 
-#copy the init script inside rootfs/init
+# Revised /init script
 cat <<EOF > $ROOTFS/init
 #!/bin/sh
 mount -t proc proc /proc
 mount -t sysfs sysfs /sys
 mount -t devtmpfs devtmpfs /dev
 
-# Mount the root project dir to /src
-mount -t 9p -o trans=virtio,version=9p2000.L project_root /src
+# --- Mount only your source/projects directory ---
+# 'projectshare' is the tag we will use in the QEMU command
+mount -t virtiofs projectshare /src
 
-echo "--- RISC-V RESEARCH ENV READY ---"
-echo "Project root is shared at /src"
-cd /src
-/bin/sh
+echo "------------------------------------------------"
+echo "NVMe Research Environment Loaded"
+echo "Shared projects are available in /src"
+echo "------------------------------------------------"
+
+cd src/
+
+exec /bin/sh
 EOF
 
 chmod +x $ROOTFS/init
 
-echo "--- 4. Packing thirdparties/rootfs.cpio ---"
+echo "--- 4. Packing rootfs  ---"
+
+# Cpio packing, pack the Busybox
 cd $ROOTFS
-find . | cpio -H newc -o > $TP_DIR/../rootfs.cpio
+find . -mindepth 1 | cpio -H newc -o | gzip -9 > $TP_DIR/../rootfs.cpio.gz
 cd ../../../
 
 echo "--- 5. Compile kernel on RISC-V target ---"
 
-cd thirdparties/permanent/linux
-make ARCH=riscv CROSS_COMPILE=riscv64-linux-gnu- menuconfig
-make ARCH=riscv CROSS_COMPILE=riscv64-linux-gnu- -j$(nproc)
-riscv64-linux-gnu-objcopy -O binary vmlinux Image
-mv Image ../../
-cd ../../../
+
+IMAGE_PATH="thirdparties/Image"
+if [ ! -f "$IMAGE_PATH" ]; then
+    echo "Kernel Image not found. Building..."
+    cd thirdparties/permanent/linux
+    make ARCH=riscv CROSS_COMPILE=$CROSS -j$(nproc)
+    riscv64-linux-gnu-objcopy -O binary vmlinux Image
+    mv Image ../../
+    cd ../../../
+else
+    echo "Kernel Image already exists."
+fi
 
 echo "--- 6. Generate NVMe disk ---"
-
 qemu-img create -f raw thirdparties/nvme_disk.img 1G
 
 echo "--- DONE ---"
